+++ sdk/confidentialledger/azure-security-confidentialledger/src/samples/java/com/azure/security/confidentialledger/PostLedgerEntrySample.java
@@ -5,6 +5,7 @@ package com.azure.security.confidentialledger;
 
 import java.io.IOException;
 
+import com.azure.core.http.HttpHeaderName;
 import org.junit.jupiter.api.Assertions;
 
 import com.azure.core.http.rest.RequestOptions;
@@ -34,7 +35,7 @@ public class PostLedgerEntrySample {
         JsonNode responseBodyJson = null;
 
         try {
-            
+
             responseBodyJson = objectMapper.readTree(parsedResponse.toBytes());
         } catch (IOException e) {
             e.printStackTrace();
@@ -44,12 +45,14 @@ public class PostLedgerEntrySample {
         String collectionId = responseBodyJson.get("collectionId").asText();
 
         // you can use the transaction id to get the transaction status
-        String transactionId = response.getHeaders().get("x-ms-ccf-transaction-id").getValue();
+        String transactionId = response.getHeaders()
+            .get(HttpHeaderName.fromString("x-ms-ccf-transaction-id"))
+            .getValue();
 
         requestOptions = new RequestOptions();
         // the transactionId can be retrieved after posting to a ledger (see PostLedgerEntry.java)
         Response<BinaryData> transactionResponse = confidentialLedgerClient.getTransactionStatusWithResponse(transactionId, requestOptions);
-    
+
         JsonNode transactionResponseBodyJson = null;
 
         try {