+++ sdk/resourcemanager/azure-resourcemanager-compute/CHANGELOG.md
@@ -4,11 +4,13 @@
 
 ### Features Added
 
-### Breaking Changes
-
-### Bugs Fixed
+- Supported updating delete options for OS disk, data disks and network interfaces attached to `VirtualMachine`.
 
 ### Other Changes
+
+#### Dependency Updates
+
+- Updated `api-version` to `2023-07-01`.
 
 ## 2.29.0 (2023-07-28)